Add toggle for off month days on static calendar
